Create default CFA layout
Create a default layout for CFA admin pages using a main header which is
much like the default main header but still adapted to CFA context.
The "+" icon in the header offers only 1 option on CFA pages, which is
"New training". It seemed not appropriate to have the options to add a
contract or such things in this part of the website, so we decided to
make it a bit different. Some changes could still be necessary but this
is already good for a start.